some refactoring for clarity
